Moves all the tests
