+++ packages/components/tests/CheckboxTest.js
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 
-import Checkbox from "../components/Checkbox";
+import Checkbox from "../Checkbox";
 
 describe( Checkbox, () => {
 	it( "matches the snapshot", () => {
