+++ packages/components/tests/HelpCenterButtonTest.js
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 
-import { HelpCenterButton } from "../components/HelpCenterButton";
+import { HelpCenterButton } from "../HelpCenterButton";
 
 test( "the HelpCenterButton matches the snapshot", () => {
 	const component = renderer.create(
