+++ packages/components/tests/IconTest.js
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 
-import Icon from "../components/Icon";
+import Icon from "../Icon";
 import { YoastSeoIcon } from "@yoast/components";
 
 test( "the Icon without props matches the snapshot", () => {
