+++ packages/components/tests/YoastWarningTest.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import YoastWarning from "../components/YoastWarning";
+import YoastWarning from "../YoastWarning";
 import { ThemeProvider } from "styled-components";
 
 test( "YoastWarning matches the snapshot", () => {
